refactor IndicatorView
